requirements.txt
aiodns==1.1.1
aiohttp==3.4.4
astroid==2.1.0
async-timeout==3.0.1
attrs==18.2.0
autopep8==1.4.3
cchardet==2.1.4
chardet==3.0.4
Click==7.0
dnspython==1.15.0
eventlet==0.24.1
Flask==1.0.2
greenlet==0.4.15
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
multidict==4.5.2
pycares==2.3.0
pycodestyle==2.4.0
pylint==2.2.2
python-engineio==3.0.0
python-socketio==2.1.0
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
yarl==1.2.6
